!175 merge feature/official-hubs-spharx01 into feature/official-hubs

Created-by: spharx01
Commit-by: spharx01;zhixianzhou
Merged-by: spharx01
Description: ## Pull Request 模板

### 变更类型 (请勾选一项)
- [ ] Bug 修复
- [ ] 新功能
- [ ] 重构/代码清理
- [ ] 性能优化
- [ ] 文档更新
- [ ] 测试补充

### 相关 Issue / 任务编号
- 大轮次编号: R-09-__
- 详细任务: R-09-__-__
- 负责团队: [A/B/C]

### 变更摘要
<!-- 简要描述本次变更的内容 -->

### 影响范围
<!-- 列出受影响的模块和文件 -->
- 模块:
- 文件:

### S-01~S-08 统一工作标准检查 (v11.14)

| # | 标准 | 检查项 | 状态 |
|---|------|--------|------|
| S-01 | 统一构建输出目录 | 构建产物在 `AgentOS-build/`，源码目录无 `.o`/`.a`/`.exe` | [ ] |
| S-02 | 统一CMake命令 | 使用 `cmake -B ../AgentOS-build` 构建 | [ ] |
| S-03 | 统一代码风格 | 所有 `.c`/`.h` 通过 `clang-format --dry-run --Werror` | [ ] |
| S-04 | 统一PR模板 | 使用本模板，填写完整 | [ ] |
| S-05 | 统一测试命令 | 使用 `cd ../AgentOS-build && ctest --output-on-failure` | [ ] |
| S-06 | 统一commit前缀 | `feat:`/`fix:`/`refactor:`/`docs:`/`test:`/`chore:` | [ ] |
| S-07 | 统一分支命名 | `feature/<name>-<issue>` / `fix/<name>-<issue>` | [ ] |
| S-08 | 统一BAN检查 | 运行 `scripts/pipeline/quality-gate.sh`，BAN-01~BAN-36 零违规 | [ ] |

### BAN 违规检查 (BAN-01~BAN-36)

- [ ] 无 BAN-01~BAN-13 违规（基础铁律）
- [ ] 无 BAN-17 违规（src/ 下无"简化实现"标记）
- [ ] 无 BAN-18 违规（src/ 下无返回固定值桩函数体）
- [ ] 无 BAN-19 违规（src/ 下无 mock/fake 降级）
- [ ] 无 BAN-20 违规（解析器无功能缺失）
- [ ] 无 BAN-25 违规（二进制无明文 License 字符串）
- [ ] 无 BAN-26 违规（Trial 版未正确限制）
- [ ] 无 BAN-28~BAN-29 违规（声明即实现，无桩函数）
- [ ] 无 BAN-30~BAN-31 违规（无资源泄漏，getter无副作用）
- [ ] 无 BAN-32 违规（无硬编码平台路径）
- [ ] 无 BAN-33 违规（构建产物不在源码目录内）
- [ ] 无 BAN-34 违规（代码已 clang-format 格式化）
- [ ] 无 BAN-35 违规（CI 无硬编码路径）
- [ ] 无 BAN-36 违规（无私有构建脚本）
- [ ] 无 TODO/FIXME/HACK 标记

### 构建验证命令（符合 v11.14 BAN-33 规范）
```bash
# ✅ 正确：构建产物在 AgentOS-build/ 目录（源码目录保持纯净）
cd AgentOS
cmake -B ../AgentOS-build -DBUILD_TESTS=ON -DCMAKE_BUILD_TYPE=Release
cmake --build ../AgentOS-build --parallel 8
cd ../AgentOS-build && ctest --output-on-failure

# ❌ 禁止：以下方式违反 BAN-33（在源码目录内构建）
# cmake -B build
# cmake --build .
```

### 跨平台验证
- [ ] Linux GCC 编译通过
- [ ] Windows MSVC 编译通过（如适用）
- [ ] 无 CROSS-01 违规（无 pthread.h/clock_gettime 直接使用）

### Commit Message 格式规范
```
<类型>(<范围>): <简短描述>

<body>

<footer>
```
**类型**: feat/fix/refactor/docs/test/chore
**范围**: checkpoint/multiagent/memoryrovol/coreloopthree/daemon/commons

### 审查清单
- [ ] 代码符合 .clang-format 规范
- [ ] 所有公共 API 有 Doxygen 注释
- [ ] 错误路径正确处理（非桩函数返回）
- [ ] 内存分配失败有安全回退
- [ ] 跨平台兼容（platform.h 抽象层）

### 额外说明
<!-- 其他需要审查者注意的事项 -->


See merge request: openairymax/agentos!176

+++ agentos/atoms/coreloopthree/src/cognition/entity_extractor.c
@@ -5,73 +5,96 @@
  */
 
 #include "entity_extractor.h"
+
 #include "intent_utils.h"
-#include <string.h>
+#include "memory_compat.h"
+
 #include <ctype.h>
-#include <stdlib.h>
 #include <regex.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
 
 static int g_extractor_initialized = 0;
 
-const char* agentos_entity_type_name(agentos_entity_type_t type) {
+const char *agentos_entity_type_name(agentos_entity_type_t type)
+{
     switch (type) {
-        case AGENTOS_ENTITY_UNKNOWN:      return "unknown";
-        case AGENTOS_ENTITY_PERSON:       return "person";
-        case AGENTOS_ENTITY_ORGANIZATION: return "organization";
-        case AGENTOS_ENTITY_LOCATION:     return "location";
-        case AGENTOS_ENTITY_TIME:         return "time";
-        case AGENTOS_ENTITY_DATE:         return "date";
-        case AGENTOS_ENTITY_NUMBER:       return "number";
-        case AGENTOS_ENTITY_URL:          return "url";
-        case AGENTOS_ENTITY_EMAIL:        return "email";
-        case AGENTOS_ENTITY_FILEPATH:     return "filepath";
-        case AGENTOS_ENTITY_COMMAND:      return "command";
-        case AGENTOS_ENTITY_PARAMETER:    return "parameter";
-        default:                          return "unknown";
+    case AGENTOS_ENTITY_UNKNOWN:
+        return "unknown";
+    case AGENTOS_ENTITY_PERSON:
+        return "person";
+    case AGENTOS_ENTITY_ORGANIZATION:
+        return "organization";
+    case AGENTOS_ENTITY_LOCATION:
+        return "location";
+    case AGENTOS_ENTITY_TIME:
+        return "time";
+    case AGENTOS_ENTITY_DATE:
+        return "date";
+    case AGENTOS_ENTITY_NUMBER:
+        return "number";
+    case AGENTOS_ENTITY_URL:
+        return "url";
+    case AGENTOS_ENTITY_EMAIL:
+        return "email";
+    case AGENTOS_ENTITY_FILEPATH:
+        return "filepath";
+    case AGENTOS_ENTITY_COMMAND:
+        return "command";
+    case AGENTOS_ENTITY_PARAMETER:
+        return "parameter";
+    default:
+        return "unknown";
     }
 }
 
-int agentos_entity_extractor_init(void) {
+int agentos_entity_extractor_init(void)
+{
     if (g_extractor_initialized) {
         return 0;
     }
-    
+
     g_extractor_initialized = 1;
     return 0;
 }
 
-void agentos_entity_extractor_cleanup(void) {
+void agentos_entity_extractor_cleanup(void)
+{
     g_extractor_initialized = 0;
 }
 
-agentos_extraction_result_t* agentos_extraction_result_create(size_t initial_capacity) {
+agentos_extraction_result_t *agentos_extraction_result_create(size_t initial_capacity)
+{
     if (initial_capacity == 0) {
         initial_capacity = 10;
     }
-    
-    agentos_extraction_result_t* result =
-        (agentos_extraction_result_t*)AGENTOS_CALLOC(1, sizeof(agentos_extraction_result_t));
+
+    agentos_extraction_result_t *result =
+        (agentos_extraction_result_t *)AGENTOS_CALLOC(1, sizeof(agentos_extraction_result_t));
     if (!result) {
         return NULL;
     }
-    
-    result->entities = (agentos_entity_t*)AGENTOS_CALLOC(initial_capacity, sizeof(agentos_entity_t));
+
+    result->entities =
+        (agentos_entity_t *)AGENTOS_CALLOC(initial_capacity, sizeof(agentos_entity_t));
     if (!result->entities) {
         AGENTOS_FREE(result);
         return NULL;
     }
-    
+
     result->entity_count = 0;
     result->capacity = initial_capacity;
-    
+
     return result;
 }
 
-void agentos_extraction_result_destroy(agentos_extraction_result_t* result) {
+void agentos_extraction_result_destroy(agentos_extraction_result_t *result)
+{
     if (!result) {
         return;
     }
-    
+
     if (result->entities) {
         for (size_t i = 0; i < result->entity_count; i++) {
             if (result->entities[i].value) {
@@ -80,43 +103,44 @@ void agentos_extraction_result_destroy(agentos_extraction_result_t* result) {
         }
         AGENTOS_FREE(result->entities);
     }
-    
+
     AGENTOS_FREE(result);
 }
 
-int agentos_extraction_result_add(agentos_extraction_result_t* result,
-                                  const agentos_entity_t* entity) {
+int agentos_extraction_result_add(agentos_extraction_result_t *result,
+                                  const agentos_entity_t *entity)
+{
     if (!result || !entity) {
         return -1;
     }
-    
+
     /* 扩容检查 */
     if (result->entity_count >= result->capacity) {
         size_t new_capacity = result->capacity * 2;
-        agentos_entity_t* new_entities = (agentos_entity_t*)AGENTOS_REALLOC(
+        agentos_entity_t *new_entities = (agentos_entity_t *)AGENTOS_REALLOC(
             result->entities, new_capacity * sizeof(agentos_entity_t));
         if (!new_entities) {
             return -1;
         }
-        
+
         /* 初始化新空间 */
         memset(new_entities + result->capacity, 0,
                (new_capacity - result->capacity) * sizeof(agentos_entity_t));
-        
+
         result->entities = new_entities;
         result->capacity = new_capacity;
     }
-    
+
     /* 复制实体 */
-    agentos_entity_t* target = &result->entities[result->entity_count];
+    agentos_entity_t *target = &result->entities[result->entity_count];
     target->type = entity->type;
     target->type_name = entity->type_name;
     target->start_pos = entity->start_pos;
     target->end_pos = entity->end_pos;
     target->confidence = entity->confidence;
-    
+
     if (entity->value && entity->value_len > 0) {
-        target->value = (char*)AGENTOS_MALLOC(entity->value_len + 1);
+        target->value = (char *)AGENTOS_MALLOC(entity->value_len + 1);
         if (target->value) {
             memcpy(target->value, entity->value, entity->value_len);
             target->value[entity->value_len] = '\0';
@@ -129,35 +153,36 @@ int agentos_extraction_result_add(agentos_extraction_result_t* result,
         target->value = NULL;
         target->value_len = 0;
     }
-    
+
     result->entity_count++;
     return 0;
 }
 
 /* 提取数字实体 */
-static void extract_numbers(const char* input, size_t input_len,
-                            agentos_extraction_result_t* result) {
-    const char* p = input;
-    int pos = 0;
-    
+static void extract_numbers(const char *input, size_t input_len,
+                            agentos_extraction_result_t *result)
+{
+    const char *p = input;
+    size_t pos = 0;
+
     while (*p && pos < input_len) {
         if (isdigit(*p)) {
-            const char* start = p;
+            const char *start = p;
             int start_pos = pos;
-            
+
             /* 提取完整数字（包括小数点） */
             while (*p && (isdigit(*p) || *p == '.')) {
                 p++;
                 pos++;
             }
-            
+
             size_t len = p - start;
             if (len > 0 && len <= 20) { /* 合理的数字长度限制 */
                 agentos_entity_t entity;
                 memset(&entity, 0, sizeof(entity));
                 entity.type = AGENTOS_ENTITY_NUMBER;
                 entity.type_name = agentos_entity_type_name(AGENTOS_ENTITY_NUMBER);
-                entity.value = (char*)AGENTOS_MALLOC(len + 1);
+                entity.value = (char *)AGENTOS_MALLOC(len + 1);
                 if (entity.value) {
                     memcpy(entity.value, start, len);
                     entity.value[len] = '\0';
@@ -165,34 +190,34 @@ static void extract_numbers(const char* input, size_t input_len,
                     entity.start_pos = start_pos;
                     entity.end_pos = pos - 1;
                     entity.confidence = 0.95f;
-                    
+
                     agentos_extraction_result_add(result, &entity);
                     AGENTOS_FREE(entity.value); /* add 已复制 */
                 }
             }
-            
+
             continue;
         }
-        
+
         p++;
         pos++;
     }
 }
 
 /* 提取 URL 实体 */
-static void extract_urls(const char* input, size_t input_len,
-                         agentos_extraction_result_t* result) {
+static void extract_urls(const char *input, size_t input_len, agentos_extraction_result_t *result)
+{
     regex_t regex;
     regmatch_t match;
-    const char* pattern = "(https?://[^\\s]+|ftp://[^\\s]+)";
-    
+    const char *pattern = "(https?://[^\\s]+|ftp://[^\\s]+)";
+
     if (regcomp(&regex, pattern, REG_EXTENDED) != 0) {
         return;
     }
-    
-    const char* p = input;
+
+    const char *p = input;
     int offset = 0;
-    
+
     while (regexec(&regex, p, 1, &match, 0) == 0) {
         size_t len = match.rm_eo - match.rm_so;
         if (len > 3 && len <= 256) { /* 合理的 URL 长度限制 */
@@ -200,7 +225,7 @@ static void extract_urls(const char* input, size_t input_len,
             memset(&entity, 0, sizeof(entity));
             entity.type = AGENTOS_ENTITY_URL;
             entity.type_name = agentos_entity_type_name(AGENTOS_ENTITY_URL);
-            entity.value = (char*)AGENTOS_MALLOC(len + 1);
+            entity.value = (char *)AGENTOS_MALLOC(len + 1);
             if (entity.value) {
                 memcpy(entity.value, p + match.rm_so, len);
                 entity.value[len] = '\0';
@@ -208,33 +233,33 @@ static void extract_urls(const char* input, size_t input_len,
                 entity.start_pos = offset + match.rm_so;
                 entity.end_pos = offset + match.rm_eo - 1;
                 entity.confidence = 0.98f;
-                
+
                 agentos_extraction_result_add(result, &entity);
                 AGENTOS_FREE(entity.value);
             }
         }
-        
+
         p += match.rm_eo;
         offset += match.rm_eo;
     }
-    
+
     regfree(&regex);
 }
 
 /* 提取邮箱实体 */
-static void extract_emails(const char* input, size_t input_len,
-                           agentos_extraction_result_t* result) {
+static void extract_emails(const char *input, size_t input_len, agentos_extraction_result_t *result)
+{
     regex_t regex;
     regmatch_t match;
-    const char* pattern = "[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}";
-    
+    const char *pattern = "[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}";
+
     if (regcomp(&regex, pattern, REG_EXTENDED) != 0) {
         return;
     }
-    
-    const char* p = input;
+
+    const char *p = input;
     int offset = 0;
-    
+
     while (regexec(&regex, p, 1, &match, 0) == 0) {
         size_t len = match.rm_eo - match.rm_so;
         if (len > 5 && len <= 128) { /* 合理的邮箱长度限制 */
@@ -242,7 +267,7 @@ static void extract_emails(const char* input, size_t input_len,
             memset(&entity, 0, sizeof(entity));
             entity.type = AGENTOS_ENTITY_EMAIL;
             entity.type_name = agentos_entity_type_name(AGENTOS_ENTITY_EMAIL);
-            entity.value = (char*)AGENTOS_MALLOC(len + 1);
+            entity.value = (char *)AGENTOS_MALLOC(len + 1);
             if (entity.value) {
                 memcpy(entity.value, p + match.rm_so, len);
                 entity.value[len] = '\0';
@@ -250,33 +275,34 @@ static void extract_emails(const char* input, size_t input_len,
                 entity.start_pos = offset + match.rm_so;
                 entity.end_pos = offset + match.rm_eo - 1;
                 entity.confidence = 0.97f;
-                
+
                 agentos_extraction_result_add(result, &entity);
                 AGENTOS_FREE(entity.value);
             }
         }
-        
+
         p += match.rm_eo;
         offset += match.rm_eo;
     }
-    
+
     regfree(&regex);
 }
 
 /* 提取文件路径实体 */
-static void extract_filepaths(const char* input, size_t input_len,
-                              agentos_extraction_result_t* result) {
+static void extract_filepaths(const char *input, size_t input_len,
+                              agentos_extraction_result_t *result)
+{
     regex_t regex;
     regmatch_t match;
-    const char* pattern = "(/[a-zA-Z0-9_./-]+|[A-Za-z]:\\\\[a-zA-Z0-9_./\\\\-]+)";
-    
+    const char *pattern = "(/[a-zA-Z0-9_./-]+|[A-Za-z]:\\\\[a-zA-Z0-9_./\\\\-]+)";
+
     if (regcomp(&regex, pattern, REG_EXTENDED) != 0) {
         return;
     }
-    
-    const char* p = input;
+
+    const char *p = input;
     int offset = 0;
-    
+
     while (regexec(&regex, p, 1, &match, 0) == 0) {
         size_t len = match.rm_eo - match.rm_so;
         if (len > 2 && len <= 512) { /* 合理的路径长度限制 */
@@ -284,7 +310,7 @@ static void extract_filepaths(const char* input, size_t input_len,
             memset(&entity, 0, sizeof(entity));
             entity.type = AGENTOS_ENTITY_FILEPATH;
             entity.type_name = agentos_entity_type_name(AGENTOS_ENTITY_FILEPATH);
-            entity.value = (char*)AGENTOS_MALLOC(len + 1);
+            entity.value = (char *)AGENTOS_MALLOC(len + 1);
             if (entity.value) {
                 memcpy(entity.value, p + match.rm_so, len);
                 entity.value[len] = '\0';
@@ -292,44 +318,44 @@ static void extract_filepaths(const char* input, size_t input_len,
                 entity.start_pos = offset + match.rm_so;
                 entity.end_pos = offset + match.rm_eo - 1;
                 entity.confidence = 0.92f;
-                
+
                 agentos_extraction_result_add(result, &entity);
                 AGENTOS_FREE(entity.value);
             }
         }
-        
+
         p += match.rm_eo;
         offset += match.rm_eo;
     }
-    
+
     regfree(&regex);
 }
 
-int agentos_entity_extract(const char* input, size_t input_len,
-                          agentos_extraction_result_t* result) {
+int agentos_entity_extract(const char *input, size_t input_len, agentos_extraction_result_t *result)
+{
     if (!input || !result || input_len == 0) {
         return -1;
     }
-    
+
     if (!g_extractor_initialized) {
         agentos_entity_extractor_init();
     }
-    
+
     /* 如果结果为空，创建新的 */
     if (!result->entities) {
-        agentos_extraction_result_t* new_result = agentos_extraction_result_create(10);
+        agentos_extraction_result_t *new_result = agentos_extraction_result_create(10);
         if (!new_result) {
             return -1;
         }
         memcpy(result, new_result, sizeof(*result));
         AGENTOS_FREE(new_result);
     }
-    
+
     /* 按类型提取各类实体 */
     extract_numbers(input, input_len, result);
     extract_urls(input, input_len, result);
     extract_emails(input, input_len, result);
     extract_filepaths(input, input_len, result);
-    
+
     return 0;
 }
